refactor: move pendingWithdrawals to separate hook used in AccountProvider
