+++ src/features/unstake/useUnstaking.ts
@@ -1,15 +1,14 @@
-import { useCallback, useEffect, useState } from 'react';
+import { useCallback } from 'react';
 import { GAS_LIMIT, GAS_PRICE } from 'src/config/consts';
 import { fromStCeloWei, toStCeloWei } from 'src/formatters/amount';
 import { useContracts } from 'src/hooks/useContracts';
 import { useAccountContext } from 'src/providers/AccountProvider';
 import { useExchangeContext } from 'src/providers/ExchangeProvider';
-import { Celo, StCelo, StCeloWei } from 'src/types/units';
-import { PendingWithdrawal } from './types';
+import { StCelo, StCeloWei } from 'src/types/units';
 
 export function useUnstaking() {
   const { address, loadBalances } = useAccountContext();
-  const { managerContract, accountContract } = useContracts();
+  const { managerContract } = useContracts();
   const { stCeloExchangeRate } = useExchangeContext();
 
   const createTxOptions = useCallback(
@@ -51,30 +50,10 @@ export function useUnstaking() {
     [stCeloExchangeRate]
   );
 
-  const [pendingWithdrawals, setPendingWithdrawals] = useState<PendingWithdrawal[]>([]);
-  const loadPendingWithdrawals = useCallback(async () => {
-    const { values = [], timestamps = [] } =
-      (await accountContract.methods.getPendingWithdrawals(address).call({ from: address })) || {};
-
-    setPendingWithdrawals(
-      values.map((amount: string, index: number) => ({
-        amount: new Celo(amount),
-        timestamp: timestamps[index],
-      }))
-    );
-  }, [accountContract, address]);
-
-  useEffect(() => {
-    // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    loadPendingWithdrawals();
-  }, [loadPendingWithdrawals]);
-
   return {
     unstake,
     stCeloExchangeRate,
     estimateUnstakingFee,
     estimateWithdrawalValue,
-    pendingWithdrawals,
-    loadPendingWithdrawals,
   };
 }
