+++ src/types/account.ts
@@ -1,0 +1,6 @@
+import { Celo } from 'src/types/units';
+
+export interface PendingWithdrawal {
+  amount: Celo;
+  timestamp: string;
+}
